+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 Unreleased
 ----------
-feat(Instrument.trade_history): add new param: ``all_=False``
+* feat(Instrument.trade_history): add new param: ``all_=False``
+* feat(Instrument.intraday_closing_price): new method
 
 v0.40.0
 -------